--- a/cpp/benchmarks/README.md
+++ b/cpp/benchmarks/README.md
@@ -1,0 +1,3 @@
+# Microbenchmarks
+
+See c/benchmarks for instructions