--- a/cpp/benchmarks/benchmarks_main.cpp
+++ b/cpp/benchmarks/benchmarks_main.cpp
@@ -1,0 +1,3 @@
+#include <benchmark/benchmark.h>
+
+BENCHMARK_MAIN();