@jdanekrh 2019-11-22
Found a Jira for this, https://issues.apache.org/jira/browse/PROTON-220 ;)