@jdanekrh 2019-11-22
I'll rename the directory.

The library is C++. It is C++ 11 in that the recommended usage is with the `for (auto _ : state) {` loop. It has 03 variant for it too, but that is supposed to have bigger overhead.