@jdanekrh 2019-11-25
Initial set of tests
```
$ cmake .. -DENABLE_BENCHMARKS=ON
$ ctest -V -R benchmarks
```
<details><summary>Expand</summary>
<p>
    
    UpdateCTestConfiguration  from :/home/jdanek/repos/qpid/qpid-proton/cmake-build-release/DartConfiguration.tcl
    Parse Config file:/home/jdanek/repos/qpid/qpid-proton/cmake-build-release/DartConfiguration.tcl
    UpdateCTestConfiguration  from :/home/jdanek/repos/qpid/qpid-proton/cmake-build-release/DartConfiguration.tcl
    Parse Config file:/home/jdanek/repos/qpid/qpid-proton/cmake-build-release/DartConfiguration.tcl
    Test project /home/jdanek/repos/qpid/qpid-proton/cmake-build-release
    Constructing a list of tests
    Done constructing a list of tests
    Updating test list for fixtures
    Added 0 tests to meet fixture requirements
    Checking test dependency graph...
    Checking test dependency graph end
    test 1
        Start  1: c-benchmarks

    1: Test command: /home/jdanek/repos/qpid/qpid-proton/cmake-build-release/c/benchmarks/c-benchmarks
    1: Test timeout computed to be: 1500
    1: 2019-11-25 14:56:51
    1: Running /home/jdanek/repos/qpid/qpid-proton/cmake-build-release/c/benchmarks/c-benchmarks
    1: Run on (8 X 4200 MHz CPU s)
    1: CPU Caches:
    1:   L1 Data 32K (x4)
    1:   L1 Instruction 32K (x4)
    1:   L2 Unified 256K (x4)
    1:   L3 Unified 8192K (x1)
    1: Load Average: 1.03, 0.84, 0.81
    1: -------------------------------------------------------------------------------------
    1: Benchmark                                           Time             CPU   Iterations
    1: -------------------------------------------------------------------------------------
    1: BM_InitCloseSender                               6.56 us         6.56 us       103901
    1: BM_InitCloseReceiver                             6.59 us         6.59 us       110445
    1: BM_EstablishConnection                           58.6 us         58.6 us        12014 items_per_second=1.42015/s messages
    1: BM_SendReceiveMessages/creditWindow:1           0.007 ms        0.007 ms       102669 items_per_second=49.1243k/s messages
    1: BM_SendReceiveMessages/creditWindow:3           0.016 ms        0.016 ms        44874 items_per_second=64.4187k/s messages
    1: BM_SendReceiveMessages/creditWindow:9           0.039 ms        0.039 ms        17263 items_per_second=77.4869k/s messages
    1: BM_SendReceiveMessages/creditWindow:27          0.118 ms        0.118 ms         5930 items_per_second=76.3999k/s messages
    1: BM_SendReceiveMessages/creditWindow:81          0.342 ms        0.342 ms         2054 items_per_second=78.8127k/s messages
    1: BM_SendReceiveMessages/creditWindow:243         0.993 ms        0.993 ms          731 items_per_second=81.0102k/s messages
    1: BM_SendReceiveMessages/creditWindow:729          2.50 ms         2.50 ms          287 items_per_second=80.0469k/s messages
    1: BM_SendReceiveMessages/creditWindow:2187         2.47 ms         2.47 ms          284 items_per_second=80.9756k/s messages
    1: BM_SendReceiveMessages/creditWindow:6561         2.57 ms         2.57 ms          266 items_per_second=77.6695k/s messages
    1: BM_SendReceiveMessages/creditWindow:19683        2.84 ms         2.84 ms          203 items_per_second=69.8212k/s messages
    1: BM_SendReceiveMessages/creditWindow:59049        4.65 ms         4.65 ms          109 items_per_second=41.529k/s messages
    1: BM_SendReceiveMessages/creditWindow:177147       81.3 ms         81.3 ms           10 items_per_second=1022.63/s messages
    1: BM_SendReceiveMessages/creditWindow:200000       87.1 ms         87.1 ms           10 items_per_second=953.899/s messages
    1: BM_SendReceiveMessages/creditWindow:1000         2.46 ms         2.46 ms          292 items_per_second=81.3426k/s messages
    1: BM_EncodeListMessage/entries:10                 0.003 ms        0.003 ms       243715 items_per_second=3.48287M/s entries
    1: BM_EncodeListMessage/entries:100                0.028 ms        0.028 ms        25122 items_per_second=3.5689M/s entries
    1: BM_EncodeListMessage/entries:10000               6.48 ms         6.48 ms          107 items_per_second=1.54238M/s entries
    1: BM_CreateListMessage                             18.7 ns         18.7 ns     36550790 items_per_second=107.191M/s entries
    1: BM_EncodeMapMessage/entries:10                  0.004 ms        0.004 ms       181525 items_per_second=2.58808M/s entries
    1: BM_EncodeMapMessage/entries:100                 0.064 ms        0.064 ms        10869 items_per_second=1.56887M/s entries
    1: BM_EncodeMapMessage/entries:10000                14.2 ms         14.2 ms           49 items_per_second=705.937k/s entries
    1: BM_CreateMapMessage                              18.2 ns         18.2 ns     37968841 items_per_second=54.8573M/s entries
    1/2 Test  #1: c-benchmarks .....................   Passed   24.88 sec
    test 26
        Start 26: cpp-benchmarks

    26: Test command: /home/jdanek/repos/qpid/qpid-proton/cmake-build-release/cpp/benchmarks/cpp-benchmarks
    26: Test timeout computed to be: 1500
    26: 2019-11-25 14:57:16
    26: Running /home/jdanek/repos/qpid/qpid-proton/cmake-build-release/cpp/benchmarks/cpp-benchmarks
    26: Run on (8 X 4200 MHz CPU s)
    26: CPU Caches:
    26:   L1 Data 32K (x4)
    26:   L1 Instruction 32K (x4)
    26:   L2 Unified 256K (x4)
    26:   L3 Unified 8192K (x1)
    26: Load Average: 1.07, 0.87, 0.82
    26: ----------------------------------------------------------------------
    26: Benchmark                            Time             CPU   Iterations
    26: ----------------------------------------------------------------------
    26: BM_UnsuccessfulContainerRun        163 us          162 us         4299
    2/2 Test #26: cpp-benchmarks ...................   Passed    0.90 sec

    The following tests passed:
            c-benchmarks
            cpp-benchmarks

    100% tests passed, 0 tests failed out of 2

    Total Test time (real) =  25.80 sec
</p>
</details>

Without the https://github.com/apache/qpid-proton/pull/210 fix

    26: BM_UnsuccessfulContainerRun      43473 us        43473 us           16

with the fix

    26: BM_UnsuccessfulContainerRun        163 us          162 us         4299